Extract lightning wallet creation from wallet list
